requirements.txt
asgiref==3.5.2
Django==4.1.1
psycopg==3.1.1
psycopg2==2.9.3
sqlparse==0.4.2
typing_extensions==4.3.0
djangorestframework==3.13.1
ipython